+++ src/app/product/services/products-service.ts
@@ -1,5 +1,7 @@
 import { Injectable } from '@angular/core';
 
+import { Observable, of } from 'rxjs';
+
 import { Category } from '../../shared/models/category.enum';
 import { Product } from '../../shared/models/product';
 
@@ -7,8 +9,8 @@ import { Product } from '../../shared/models/product';
   providedIn: 'root'
 })
 export class ProductsService {
-  getProducts(): Product[] {
-    return [
+  getProducts(): Observable<Product[]> {
+    return of([
       {
         name: 'Snowman', description: 'Ice-cream Snowman', price: 5.00, category: Category.Food, isAvailable: true,
       },
@@ -18,6 +20,6 @@ export class ProductsService {
       {
         name: 'BigFoot', description: 'Boots', price: 50.00, category: Category.Clothing, isAvailable: true,
       },
-    ];
+    ]);
   }
 }