+++ src/app/shared/shared.module.ts
@@ -3,11 +3,11 @@ import { CommonModule } from '@angular/common';
 
 import { HighlightDirective } from './directives/highlight.directive';
 import { ClickHandlerDirective } from './directives/click-handler.directive';
-
+import { OrderByPipe } from './pipes/order-by.pipe';
 
 @NgModule({
-  declarations: [HighlightDirective, ClickHandlerDirective],
-  exports: [HighlightDirective, ClickHandlerDirective],
+  declarations: [HighlightDirective, ClickHandlerDirective, OrderByPipe],
+  exports: [HighlightDirective, ClickHandlerDirective, OrderByPipe],
   imports: [
     CommonModule
   ]